code reorganization with module
