+++ index.html
@@ -15,8 +15,9 @@
     <main id="main">
         <div class="photographer_section"></div>
     </main>
-    <script src="scripts/factories/photographerFactory.js"></script>
-    <script src="scripts/pages/index.js"></script>
+    <!-- <script type="module" src="scripts/main.js"></script> -->
+    <!-- <script src="scripts/factories/photographerFactory.js"></script> -->
+    <script type="module" src="scripts/pages/index.js"></script>
     <!-- <script src="scripts/pages/photographer.js"></script> -->
   </body>
 </html>
