+++ photographer.html
@@ -44,13 +44,14 @@
         </div>
       </div>
     </main>
-    <script src="scripts/factories/photographerFactory.js"></script>
-    <script src="scripts/factories/mediaFactory.js"></script>
-    <script src="scripts/factories/modalFactory.js"></script>
-    <script src="scripts/pages/photographer.js"></script>
-    <script src="scripts/pages/media.js"></script>
+    <!-- <script type="module" src="scripts/main.js"></script> -->
+    <!-- <script src="scripts/factories/photographerFactory.js"></script>
+    <script src="scripts/factories/mediaFactory.js"></script> -->
+    <!-- <script type="module"   src="scripts/factories/modalFactory.js"></script> -->
+    <script type="module" src="scripts/pages/photographer.js"></script>
+    <!-- <script type="module" src="scripts/pages/media.js"></script> -->
     <script src="scripts/utils/dropDown.js"></script>
-    <script src="/scripts/utils/contactForm.js"></script>
+    <!-- <script type="module" src="/scripts/utils/contactForm.js"></script> -->
     <script src="scripts/utils/lightbox.js"></script>
   </body>
 </html>
