+++ scripts/controllers/idPhotographer.js
@@ -1,0 +1,9 @@
+// // Récupération de l'id du photographe
+// export function getPhotographerId() {
+//     const queryString = window.location.search;
+//     const urlParams = new URLSearchParams(queryString);
+//     const idPhotographer = urlParams.get('id');
+
+//     return idPhotographer
+// }
+
