+++ scripts/factories/photographerFactory.js
@@ -1,4 +1,6 @@
-function photographerFactory(data) {
+import { displayContactModal } from "../utils/contactForm.js"
+
+export function photographerFactory(data) {
     const { name, id, city, country, tagline, price, portrait } = data;
     const picture = `assets/photographers/IDPhotos/${portrait}`;
     const place = `${city}, ${country}`;
