+++ scripts/utils/contactForm.js
@@ -1,13 +1,32 @@
-// récupération des datas medias
-async function getMedias() {
-  const mediasData = "../../data/photographers.json";
-  const response = await fetch(mediasData);
-  const dataMedias = await response.json();
-  const medias = dataMedias;
+import { getMedias } from "../database/services.js"
+import { modalFactory } from "../factories/modalFactory.js"
+import photographerName from "../pages/photographer.js"
 
-  return medias
+// récupération des éléments html
+const modal = document.querySelector(".modal");
+const contactModal = document.querySelector(".contact_container");
+const submit = document.querySelector("input[type=submit]");
+
+// affichage de la modale de contact
+export function displayContactModal() {
+  modal.style.display = "block";
+  contactModal.style.display = "block";
+  lightboxModal.style.display = "none";
 }
 
+// fermeture de la modale
+export function closeContactModal() {
+  modal.style.display = "none";
+}
+
+// enlève le formulaire de la modale
+function hideForm() {
+  const form = document.querySelector("form");
+  form.style.display = "none";
+}
+
+// affichage des datas du photographe sélectionné dans le header
+// via la modalFactory
 async function displayDataInContactForm(medias) {
   const contactSection = document.querySelector(".contact_container");
   medias.photographerName = photographerName;
@@ -23,53 +42,30 @@ async function init() {
 
 init();
 
-// récupération des éléments html
-const modal = document.querySelector(".modal");
-const contactModal = document.querySelector(".contact_container");
-const submit = document.querySelector("input[type=submit]");
-
-// affichage de la modale de contact
-function displayContactModal() {
-  modal.style.display = "block";
-  contactModal.style.display = "block";
-  lightboxModal.style.display = "none";
-}
-
-// // fermeture de la modale
-function closeContactModal() {
-  modal.style.display = "none";
-}
-
-// // enlève le formulaire de la modale
-function hideForm() {
-  const form = document.querySelector("form");
-  form.style.display = "none";
-}
-
-// // affiche le message de validation dans la modale
+// affiche le message de validation dans la modale
 function displayValidation() {
   const validation = document.querySelector(".validation");
   validation.style.display = "flex";
 }
 
-// // check si l'input prénom est valide et renvoi si message erreur ou non
-function isFirstnameValid() {
+// check si l'input prénom est valide et renvoi si message erreur ou non
+export function isFirstnameValid() {
   const firstname = document.getElementById("firstname");
   let firstnameValue = firstname.value.trim();
   if (firstnameValue != "") {
     const regex = /[A-Za-z0-9]{2,}/;
     if (regex.test(firstnameValue)) {
       return hideError(firstname.parentNode);
-      } 
-  
+    } 
+    
     return showError(firstname.parentNode)
   }
-
+  
   return showError(firstname.parentNode)
 }
 
-// // check si l'input nom est valide et renvoi si message erreur ou non
-function isLastnameValid() {
+// check si l'input nom est valide et renvoi si message erreur ou non
+export function isLastnameValid() {
   const lastname = document.getElementById("lastname");
   let lastnameValue = lastname.value.trim();
   if (lastnameValue != "") {
@@ -77,15 +73,15 @@ function isLastnameValid() {
     if (regex.test(lastnameValue)) {
       return hideError(lastname.parentNode);
     } 
-
+    
     return showError(lastname.parentNode)
   }
-
+  
   return showError(lastname.parentNode)
 }
 
-// // check si l'input email est valide et renvoi si message erreur ou non
-function isEmailValid() {
+// check si l'input email est valide et renvoi si message erreur ou non
+export function isEmailValid() {
   const email = document.getElementById("email");
   let emailValue = email.value.trim();
   if (emailValue != "") {
@@ -93,15 +89,15 @@ function isEmailValid() {
     if (regex.test(emailValue)) {
       return hideError(email.parentNode);
     } 
-
+    
     return showError(email.parentNode)
   }
-
+  
   return showError(email.parentNode)
 }
 
-// // check si l'input message est valide et renvoi si message erreur ou non
-function isMessageValid() {
+// check si l'input message est valide et renvoi si message erreur ou non
+export function isMessageValid() {
   const message = document.getElementById("message");
   let messageValue = message.value.trim();
   if (messageValue != "") {
@@ -109,37 +105,37 @@ function isMessageValid() {
     if (regex.test(messageValue)) {
       return hideError(message.parentNode);
     } 
-
+    
     return showError(message.parentNode)
   }
-
+  
   return showError(message.parentNode)
 }
 
-// // affiche le message d'erreur
+// affiche le message d'erreur
 function showError(elt) {
   elt.setAttribute("data-error-visible", true);
   
   return false
 }
-  
+
 // enlève le message d'erreur
 function hideError(elt) {
   elt.removeAttribute("data-error-visible");
-
+  
   return true
 }
 
-// // check si les inputs sont valides
+// check si les inputs sont valides
 function checkInputs() {
   let isInputsValid = isFirstnameValid() && isLastnameValid() && isEmailValid() && isMessageValid();
   if (isInputsValid) {
     return true
   }
-
+  
   return false
 }
-  
+
 // // donne accès au bouton si les inputs sont valides
 function disabledSubmit() {
   const btn = document.querySelector(".contact_button");
@@ -147,21 +143,12 @@ function disabledSubmit() {
 }
 
 // // soumet le formulaire si les inputs sont valides
-function validateForm(elt) {
+export function validateForm(elt) {
   elt.preventDefault();
   if (checkInputs()) {
     hideForm();
     displayValidation();
-    // lightboxModal.style.display = "none";
     form.reset();
   }
 }
 
-// // EVENT LISTENERS
-// firstname.addEventListener("blur", isFirstnameValid);
-// lastname.addEventListener("blur", isLastnameValid);
-// email.addEventListener("blur", isEmailValid);
-// message.addEventListener("blur", isMessageValid);
-// formData.forEach(input => input.addEventListener("change", disabledSubmit));
-// form.addEventListener("submit", validateForm);
-
